pyproject.toml
[project]
name = "talkter"
version = "0.1.0"
description = ""
authors = [
    {name = "ChanHyuk Park",email = "devch1013@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "ruff (>=0.14.1,<0.15.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "firebase-admin (>=7.1.0,<8.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg[binary] (>=3.2.11,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "boto3 (>=1.40.61,<2.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "pillow (>=12.0.0,<13.0.0)"
]

[tool.poetry]
package-mode = false  # 이 줄 추가

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
